agent-eval-framework/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-eval-framework"
version = "0.1.0"
authors = [
  { name="Jules", email="jules@example.com" },
]
description = "A generic framework for evaluating AI agents with Vertex AI."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-cloud-aiplatform>=1.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
]

[project.urls]
"Homepage" = "https://github.com/example/agent-eval-framework"
"Bug Tracker" = "https://github.com/example/agent-eval-framework/issues"
